requirements.txt
docopt
requests
flask
pyyaml
toposort